refactor(resource): move create form logic to page
