+++ features/resources/containers/index.ts
@@ -1,3 +1,2 @@
-export { default as CreateResourceForm } from "./CreateResourceForm";
 export { default as ResourceReviewForm } from "./ResourceReviewForm";
 export { default as UpdateResourceForm } from "./UpdateResourceForm";
